Update MapScreen.js
conversion to unix time fixed, and time is now displayed correctly when first rendering map screen(new Date is in ms so need to take obj.time * 1000). Rendering new markers inside if statement didn't work, but we need to somehow not allow marker to be placed if it does not post to server. As of now a marker will be rendered even if it is not a successful post request